[permissions] [l] : move methods to the right class
