+++ lib/Github.js
@@ -152,55 +152,8 @@ export default class Github{
     }
   }
   
-  /**
-   * 
-   * @param {Object} repoInfo - Repository Info 
-   * @param {String} permission - permission to be filtered 
-   */
-  _filterRepoPermissions(repoInfo, permission){
-    const userObjects = repoInfo.colaborators.edges
-    const array = []
-    userObjects.filter(userObject => {
-      if(permission === 'READ'){
-        if(userObject.permission === 'READ' || 
-          userObject.permission === 'WRITE'|| 
-          userObject.permission === 'ADMIN') array.push(userObject.node.login)
-      }
-      if(permission === 'WRITE'){
-        if(userObject.permission === 'WRITE' || 
-          userObject.permission === 'ADMIN') array.push(userObject.node.login)
-      }
-      if(permission === 'ADMIN'){
-        if(userObject.permission === 'ADMIN') array.push(userObject.node.login)
-      }
-
-    })
-
-    return array
-  }
-
-  /**
-   * 
-   * @param {Object} repos - Repository List 
-   */
-  parserPermissions(repos){
-    const repoList = repos.data.organization.repositories.nodes
-
-    const parsedRepos = repoList.map(repo => {
-      const repository = {
-        dataset: repo.name,
-        readers: this._filterRepoPermissions(repo, 'READ'),
-        editors: this._filterRepoPermissions(repo, 'WRITE'),
-        admins: this._filterRepoPermissions(repo, 'ADMIN')
-      }
-
-      if(!repo.isPrivate) repository.readers.unshift('PUBLIC', 'LOGGED_IN')
-     
-      return repository
-    })
-    
-    return parsedRepos
-  }
+ 
+ 
 
 
   async getRepositoriesForUser(){
