+++ lib/Permissions.js
@@ -1,15 +1,22 @@
 import jsonwebtoken from 'jsonwebtoken'
-import datasets from '../data/organization.json'
-// import fs from 'fs'
+import { initializeApollo } from "./apolloClient"
+import { PERMISSIONS } from './queries'
+
 require('dotenv').config()
 
 export default class Permissions{
   
-  constructor(){
+  constructor(initialApolloState){
+    this.apolloClient = initializeApollo(initialApolloState)
     this.defaultOrg = process.env.NEXT_PUBLIC_ORG_NAME
-    this.datasets = datasets
+    this.reposInfo =  this.apolloClient.readQuery({ query: PERMISSIONS })
+    this.datasets
+    this.init()
   }
   
+  init(){
+    this.datasets = this._parserPermissions(this.reposInfo)
+  }
   /**
    * 
    * @param {String} dataset - Dataset name to be validated
@@ -28,12 +35,64 @@ export default class Permissions{
 
   /**
    * 
+   * @param {Object} repoInfo - Repository Info 
+   * @param {String} permission - permission to be filtered 
+   */
+  _filterRepoPermissions(repoInfo, permission){
+    
+    const array = []
+    repoInfo.collaborators.edges.filter(userObject => {
+      if(permission === 'READ'){
+        if(userObject.permission === 'READ' || 
+          userObject.permission === 'WRITE'|| 
+          userObject.permission === 'ADMIN') array.push(userObject.node.login)
+      }
+      if(permission === 'WRITE'){
+        if(userObject.permission === 'WRITE' || 
+          userObject.permission === 'ADMIN') array.push(userObject.node.login)
+      }
+      if(permission === 'ADMIN'){
+        if(userObject.permission === 'ADMIN') array.push(userObject.node.login)
+      }
+
+    })
+
+    return array
+  }
+
+  /**
+   * 
+   * @param {Object} repos - Repository List 
+   */
+  _parserPermissions(repos){
+    const repoList = repos.organization.repositories.nodes
+    const parsedRepos = repoList.map(repo => {
+      const repository = {
+        dataset: repo.name,
+        readers: this._filterRepoPermissions(repo, 'READ'),
+        editors: this._filterRepoPermissions(repo, 'WRITE'),
+        admins: this._filterRepoPermissions(repo, 'ADMIN')
+      }
+
+      if(!repo.isPrivate) repository.readers.unshift('PUBLIC', 'LOGGED_IN')
+     
+      return repository
+    })
+    
+    return parsedRepos
+  }
+
+
+
+  /**
+   * 
    * @param {String} user - username to check permission 
    * @param {String} dataset - dataset to check permission
    * @param {String} permission permission : read/write/delete
    */
-  userHasPermission(user, dataset, permission){
-    const datasets = this.datasets || []
+  async userHasPermission(user, dataset, permission){
+    
+    const datasets = this.datasets
     const repo = datasets.find(item => item.dataset === dataset)
     if(!dataset) {
       throw new Error('Dataset not found.')
@@ -47,6 +106,8 @@ export default class Permissions{
       return repo.admins.includes(user)
     }
   }
+
+
 
   /**
  *
